Add two layers of abstract classes
